Fix issue with fetching large dataset files
Prior to this, calls to beaker.Beaker.dataset.fetch that download large binary
files would hang for a long time and explode memory usage because response.text
manifests the whole file as a string internally.  Now, that will only be called
if the file is sufficiently small and not using streaming mode.